Make shutTheBox functional
Developed a working version of shut the box. This version has no u.i besides for the console. The purpose of this study will be to create an AI that can play this game, not to have a pretty version of this. Also for speed, using a console will all deep learning to occur faster